Load CNDied Webjars dependencies protocol-less
I figured out that the default configuration is not working when used in HTTPS context.

By default the dependencies are linked to an HTTP resource (eg: https://cdn.jsdelivr.net/webjars/xx/yy/zz.min.js), and the browers forbid to fetch an insecure file in HTTPS.

This PR just ensure that dependencies are always loaded in the right context.